% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl.R
\name{fibdl}
\alias{fibdl}
\title{Fonction intégratrice}
\usage{
fibdl(listflor, info_uo)
}
\arguments{
\item{listflor}{table renseignant les listes floristiques avec trois colonnes id_prelevement, taxons, et ab}

\item{info_uo}{table d'informations complémentaires sur le prélevement.}
}
\value{

}
\description{
Fonction intégratrice
}
