% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_ibdl_qualify.R
\name{fibdl_ibdl_qualify}
\alias{fibdl_ibdl_qualify}
\title{Qualification de l IBDL}
\usage{
fibdl_ibdl_qualify(data)
}
\arguments{
\item{data}{tableau de sortie de la fonction fibl_ibdl()}
}
\value{
tableau d'évaluation de la qualité écologique avec l'indice IBDL et la classe d'état, prenant en compte la bonne représentativité des données.
Plus de 75\% de la typologie des rives pris en compte dans le calcul de l'indice et au moins 3 unités d'observation échantillonnés
}
\description{
Qualification de l IBDL
}
