% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_listflor_EQR.R
\name{fibdl_listflor_EQR}
\alias{fibdl_listflor_EQR}
\title{Transformation en EQR au niveau du prélèvement}
\usage{
fibdl_listflor_EQR(data, add_information_station)
}
\arguments{
\item{add_information_station}{}
}
\value{

}
\description{
Transformation en EQR au niveau du prélèvement
}
