% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_listflor_EQR_standard.R
\name{fibdl_listflor_EQR_standard}
\alias{fibdl_listflor_EQR_standard}
\title{Standardisation des EQR au niveau du prélèvement}
\usage{
fibdl_listflor_EQR_standard(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Standardisation des EQR au niveau du prélèvement
}
