% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_listflor_metrics.R
\name{fibdl_listflor_metrics}
\alias{fibdl_listflor_metrics}
\title{Calcul des métriques au niveau du prélèvement}
\usage{
fibdl_listflor_metrics(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Calcul des métriques au niveau du prélèvement
}
