% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_listflor_transcode.R
\name{fibdl_listflor_transcode}
\alias{fibdl_listflor_transcode}
\title{Harmonisation taxonomique au niveau espèce}
\usage{
fibdl_listflor_transcode(data)
}
\arguments{
\item{data}{liste floristique dans le bon format}
}
\value{
liste floristique au niveau taxonomique espèce
}
\description{
Harmonisation taxonomique au niveau espèce
}
\examples{

}
