% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_listflor_validate.R
\name{fibdl_listflor_validate}
\alias{fibdl_listflor_validate}
\title{Validation des listes floristiques}
\usage{
fibdl_listflor_validate(data)
}
\arguments{
\item{data}{listes floristiques harmonisées via la fonction fibdl_listflor_transcode}
}
\value{
tableau des listes floristiques validées selon trois critères :
\itemize{
\item un nombre total de valves comptés supérieur à 350
\item un ratio de valves déterminées au rang taxonomique de l'espère surpérieur à 25\%
\item un ratio de valves comptées étant indiciels surpérieur à 75\%
}
}
\description{
Validation des listes floristiques
}
