% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl.R
\name{fibdl}
\alias{fibdl}
\title{Fonction integratrice pour le calcul de Indice Biologique Diatomees en Lac (IBDL)}
\usage{
fibdl(listflor, add_information_station)
}
\arguments{
\item{listflor}{tableau renseignant les listes floristiques avec trois colonnes id_prelevement, taxons, et ab}

\item{add_information_station}{tableau d informations complementaires sur le prelevement.}
}
\description{
Fonction integratrice pour le calcul de Indice Biologique Diatomees en Lac (IBDL)
}
