% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_checkof_inputs.R
\name{fibdl_checkof_inputs}
\alias{fibdl_checkof_inputs}
\title{Check of inputs}
\usage{
fibdl_checkof_inputs(list_flor, info_uo)
}
\arguments{
\item{info_uo}{}
}
\value{

}
\description{
Check of inputs
}
\examples{
list_flor <- read.csv2(system.file("listflor.csv", package = "IBDL"),fileEncoding = "utf-8")
info_uo<- utils::read.csv2(system.file("info_uo.csv", package = "IBDL"),fileEncoding = "utf-8")
fibdl_checkof_inputs(list_flor,info_uo)
}
