% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_listflor_EQR.R
\name{fibdl_listflor_EQR}
\alias{fibdl_listflor_EQR}
\title{Transformation en EQR au niveau du prélèvement}
\usage{
fibdl_listflor_EQR(data, add_information_station)
}
\arguments{
\item{data}{tableau de sortie de la fonction fibdl_listflor_metrics()}

\item{add_information_station}{tableau d'entrée renseignant sur les la nature du substrat et la classe d'alcalinité du plan d'eau échantillonné.}
}
\value{

}
\description{
Transformation en EQR au niveau du prélèvement
}
