% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_listflor_EQR_standard.R
\name{fibdl_listflor_EQR_standard}
\alias{fibdl_listflor_EQR_standard}
\title{Standardisation des EQR au niveau du prélèvement}
\usage{
fibdl_listflor_EQR_standard(data)
}
\arguments{
\item{data}{tableau de sortie de la fonction fibdl_listflor_EQR()}
}
\value{
tableau des résultats en EQR standardisés, avec des répartitions identiques et donc des seuils identiques
}
\description{
Standardisation des EQR au niveau du prélèvement
}
