% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_listflor_transcode.R
\name{fibdl_listflor_transcode}
\alias{fibdl_listflor_transcode}
\title{Title}
\usage{
fibdl_listflor_transcode(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Title
}
\examples{

}
