% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_listflor_transcode.R
\name{fibdl_listflor_transcode}
\alias{fibdl_listflor_transcode}
\title{Harmonisation taxonomique au niveau espèce}
\usage{
fibdl_listflor_transcode(data)
}
\arguments{
\item{data}{liste floristique dans le bon format}
}
\value{
liste floristique au niveau taxonomique espèce
}
\description{
Harmonisation taxonomique au niveau espèce
}
\examples{
listflor <- tibble::tibble(id_prelevement=1,
taxons=c("NITZ","ADPA","ABAS","PBTH","PHAO","PSDG","NXDS","GLAT",
"DSTE","RABB","COCE","ENCM","PSBR","PLTD","ADEU","AUPU"),
ab=c(202,20,2,26,12,5,6,74,5,2,3,3,15,10,5,20))
fibdl_listflor_transcode(listflor)

}
