% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibdl_uo_score.R
\name{fibdl_uo_score}
\alias{fibdl_uo_score}
\title{Calcul du score au niveau de l'unité d'observation}
\usage{
fibdl_uo_score(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Calcul du score au niveau de l'unité d'observation
}
